requirements.txt
flask
flask-cors
mysql-connector-python
requests
beautifulsoup4
lxml
gspread
google-auth